Refactor(auth): 로그인, 회원가입 컴포넌트 구조 수정
