+++ src/app/auth/login/_components/form/index.ts
@@ -1,0 +1,1 @@
+export { LoginForm } from './Form'
