+++ src/app/auth/signup/_components/form/field/checkbox/index.ts
@@ -1,0 +1,1 @@
+export { SignupCheckboxField } from './CheckboxField'
