+++ src/app/auth/signup/_components/form/field/index.ts
@@ -1,0 +1,2 @@
+export * from './checkbox'
+export * from './input'
