+++ src/app/auth/signup/_components/form/field/input/InputField.tsx
@@ -1,15 +1,7 @@
-import {
-  ComponentPropsWithoutRef,
-  InputHTMLAttributes,
-  useEffect,
-  useState,
-} from 'react'
+import { InputHTMLAttributes, useEffect, useState } from 'react'
 import { useFormContext } from 'react-hook-form'
 
-import * as CheckboxPrimitive from '@radix-ui/react-checkbox'
-
 import {
-  Checkbox,
   FormControl,
   FormDescription,
   FormField,
@@ -18,10 +10,10 @@ import {
   FormMessage,
   Input,
 } from '@/components/ui'
+import { cn } from '@/lib/utils'
 import { CheckRespose } from '@/service/server/auth/signup'
 
-import { CheckStudentNumberButton, CheckUserIdButton } from './CheckButton'
-import { CheckResultMessage } from './CheckResultMessage'
+import { CheckStudentNumberButton, CheckUserIdButton } from './check-button'
 
 interface SignupInputFieldProps extends InputHTMLAttributes<HTMLInputElement> {
   name: string
@@ -78,42 +70,16 @@ export const SignupInputField = ({
           </FormControl>
           <FormDescription className="pl-2">{formDescription}</FormDescription>
           <FormMessage className="pl-2" />
-          {checkResult && <CheckResultMessage checkResult={checkResult} />}
-        </FormItem>
-      )}
-    />
-  )
-}
-
-interface SignupCheckboxFieldProps
-  extends ComponentPropsWithoutRef<typeof CheckboxPrimitive.Root> {
-  name: string
-  formLabel: string
-}
-
-export const SignupCheckboxField = ({
-  name,
-  formLabel,
-}: SignupCheckboxFieldProps) => {
-  const form = useFormContext()
-
-  return (
-    <FormField
-      control={form.control}
-      name={name}
-      render={({ field }) => (
-        <FormItem>
-          <FormControl className="flex items-center gap-2 pl-2 pt-4">
-            <div>
-              <Checkbox
-                id={name}
-                checked={field.value}
-                onCheckedChange={field.onChange}
-              />
-              <FormLabel htmlFor={name}>{formLabel}</FormLabel>
-            </div>
-          </FormControl>
-          <FormMessage className="pl-3" />
+          {checkResult && (
+            <p
+              className={cn(
+                checkResult.success ? 'text-blue-600' : 'text-red-600',
+                'text-sm',
+              )}
+            >
+              {checkResult.message}
+            </p>
+          )}
         </FormItem>
       )}
     />
