+++ src/app/auth/signup/_components/form/field/input/check-button/index.ts
@@ -1,0 +1,1 @@
+export { CheckUserIdButton, CheckStudentNumberButton } from './CheckButton'
