+++ src/app/auth/signup/_components/form/field/input/index.ts
@@ -1,0 +1,1 @@
+export { SignupInputField } from './InputField'
