+++ src/app/auth/signup/_components/form/success-dialog/SuccessDialog.tsx
