+++ src/app/auth/signup/_components/form/success-dialog/index.ts
@@ -1,0 +1,1 @@
+export { SignupSuccessDialog } from './SuccessDialog'
