+++ src/app/auth/signup/_components/index.ts
@@ -1,0 +1,1 @@
+export { SignupForm } from './form'
