+++ src/app/auth/signup/page.tsx
@@ -1,19 +1,18 @@
 'use client'
 
-import { AuthCard } from '~auth/_components/AuthCard'
-import { LinkButton } from '~auth/_components/LinkButton'
+import { AuthCardLayout, LinkButton } from '@/components/feature'
 
-import { SignupForm } from './_components/SignupForm'
+import { SignupForm } from './_components/form'
 
 const SignupPage = () => {
   return (
     <div className="flex w-full flex-col items-center gap-2">
-      <AuthCard title="Sign up to HAEDAL" className="max-w-lg">
+      <AuthCardLayout title="Sign up to HAEDAL" className="max-w-lg">
         <SignupForm />
-      </AuthCard>
+      </AuthCardLayout>
       <div className="flex w-3/4 flex-col items-center justify-center pt-2 sm:flex-row">
         <div className="w-max text-sm text-input">계정이 이미 있으신가요?</div>
-        <LinkButton backLink="/auth/login" label="로그인하기" />
+        <LinkButton linkTo="/auth/login" label="로그인하기" />
       </div>
     </div>
   )
