+++ src/components/feature/auth/card-layout/index.ts
@@ -1,0 +1,1 @@
+export { AuthCardLayout } from './CardLayout'
