+++ src/components/feature/auth/link-button/LinkButton.tsx
@@ -4,14 +4,14 @@ import { Button } from '@/components/ui'
 
 type LinkButtonProps = {
   label: string
-  backLink: string
+  linkTo: string
 }
 
-export const LinkButton = ({ label, backLink }: LinkButtonProps) => {
+export const LinkButton = ({ label, linkTo }: LinkButtonProps) => {
   return (
     <div className="flex w-fit justify-center">
       <Button variant="link" className="font-normal">
-        <Link href={backLink} className="text-white">
+        <Link href={linkTo} className="text-white">
           {label}
         </Link>
       </Button>
