+++ src/components/feature/auth/link-button/index.ts
@@ -1,0 +1,1 @@
+export { LinkButton } from './LinkButton'
