+++ src/components/feature/index.ts
@@ -1,5 +1,6 @@
 export * from './activity'
 export * from './admin'
+export * from './auth'
 export * from './post'
 export * from './skeleton'
 export * from './board'
